정건희: <style> 로그인, 회원가입 타이틀, 변경 된 버튼 디자인 css 적용
