--- src/components/formSignIn/style.scss
@@ -1,148 +1,0 @@
-$mainWidth: 1180px;
-
-.background {
-  width: 100vw;
-  height: 100vh;
-}
-
-.signBox {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-
-  height: 100vh;
-
-  margin: auto;
-}
-
-.divTitleBox {
-  h2 {
-    padding-top: 50px;
-    padding-bottom: 45px;
-    color: white;
-  }
-}
-
-.divSignUpBox {
-  margin-top: 64px;
-  button {
-    border-radius: 5px;
-    width: 400px;
-    height: 50px;
-    border: none;
-    background: linear-gradient(
-      to bottom,
-      rgba(167, 167, 167, 0.47),
-      rgba(65, 65, 65, 0.9945)
-    );
-    color: white;
-    font-size: 16px;
-  }
-}
-
-.divInputBox {
-  margin-top: 25px;
-  input {
-    background-color: white;
-    width: 400px;
-    height: 50px;
-    border-radius: 5px;
-    border: none;
-    opacity: 45%;
-    font-size: 14px;
-    padding-left: 20px;
-  }
-  p {
-    margin-top: 8px;
-    font-size: 14px;
-    float: left;
-    margin-left: 50px;
-    color: #ff264a;
-    font-weight: bold;
-  }
-}
-.bodySignin {
-  width: 500px;
-  height: 600px;
-  background-color: rgba(0, 0, 0, 0.6);
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  border-radius: 10px;
-  z-index: 1;
-}
-
-.loginName {
-  font-size: 28px;
-  margin-top: -60px;
-}
-.inputBtnBox {
-  width: 400px;
-  height: 320px;
-}
-.mailInputBox {
-  width: 400px;
-  height: 90px;
-  margin-top: -25px;
-}
-
-.divButtonBox {
-  width: 400px;
-  height: 120px;
-
-  margin-top: 60px;
-}
-
-.loginBtn {
-  margin-left: 0px;
-  width: 400px;
-  height: 50px;
-  border-radius: 5px;
-  border: none;
-  background: linear-gradient(
-    to bottom,
-    rgba(167, 167, 167, 0.47),
-    rgba(65, 65, 65, 0.9945)
-  );
-  color: white;
-  font-size: 16px;
-}
-
-.googleBtn {
-  width: 400px;
-  height: 50px;
-  border-radius: 5px;
-  margin-top: 20px;
-  background-color: rgba(0, 0, 0, 0.38);
-  color: white;
-  font-size: 16px;
-  border: 0.5px solid white;
-}
-
-.divSignUpBox {
-  width: 400px;
-  height: 50px;
-
-  align-items: center;
-}
-.joinBtn {
-  font-weight: bold;
-  color: white;
-  float: left;
-  margin-left: 10px;
-}
-
-.joinText {
-  color: white;
-  opacity: 50%;
-  float: left;
-  margin-left: 40px;
-}
-.emailText {
-  display: flex;
-  gap: 5px;
-  margin-top: 10px;
-  font-size: 14px;
-  color: #ff264a;
-}
