--- src/components/googleLogin/style.scss
@@ -1,11 +1,0 @@
-.icon {
-  width: 60px;
-  float: left;
-  border-right: 2px solid white;
-}
-
-.googleLogin {
-  font-size: 16px;
-  margin-top: 3px;
-  margin-right: 50px;
-}
