+++ src/components/sign/googleLogin/style.scss
@@ -1,0 +1,17 @@
+.googleLoginBtn {
+  width: 120px;
+  height: 50px;
+
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  gap: 10px;
+
+  cursor: pointer;
+
+  .googleLogin {
+    font-size: 16px;
+    font-weight: 500;
+    color: rgba(217, 217, 217, 1);
+  }
+}
