--- src/components/temp_sign/style.scss
@@ -1,32 +1,0 @@
-.tempSign {
-  display: flex;
-  gap: 10px;
-}
-
-.signUp {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-
-  width: 200px;
-  height: 200px;
-  margin-bottom: 10px;
-  gap: 5px;
-
-  background-color: #e5e5e5;
-}
-
-.login {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-
-  width: 200px;
-  height: 200px;
-  margin-bottom: 10px;
-  gap: 5px;
-
-  background-color: #e5e5e5;
-}
