--- src/pages/SignIn.jsx
@@ -1,11 +1,0 @@
-import FormSignIn from '../components/formSignUp/FormSignUp';
-
-function SignIn() {
-  return (
-    <>
-      <FormSignIn></FormSignIn>
-    </>
-  );
-}
-
-export default SignIn;
